# Rewrite index optimization code for maximum efficiency