While all the previous fixes to optimizeVertexCache invocation fixed the vertex transform efficiency, the import code still was missing two crucial recommendations from meshoptimizer documentation:

- All meshes should be optimized for vertex fetch (this reorders vertices for maximum fetch efficiency)
- When LODs are used with a shared vertex buffer, the vertex order should be generated by doing a vertex fetch optimization on the concatenated index buffer from coarse to fine LODs; this maximizes fetch efficiency for coarse LODs

The last point is especially important for Mali GPUs; unlike other GPUs where vertex order affects fetch efficiency but not shading, these GPUs have various vertex shading quirks (depending on the GPU generation) that really require compact index ranges for each LOD, which requires the second optimization mentioned above. However all of these also help desktop GPUs and other mobile GPUs as well.

Because this optimization is "global" in the sense that it affects all LODs and all vertex arrays in concert, I've taken this opportunity to isolate all optimization code in this function and pull it out of `generate_lods` and `create_shadow_mesh`; this doesn't change the vertex cache efficiency, but makes the code cleaner. Consequently, `optimize_indices` should be called after other functions like `create_shadow_mesh` / `generate_lods`.

This required exposing `meshopt_optimizeVertexFetchRemap`; as a drive-by, `meshopt_simplifySloppy` was never used so it's not exposed anymore - this will simplify future meshopt upgrades if they end up changing the function's interface.

Future optimizations, if any, can likely be done in this function without changing any of the external code; for example, input mesh geometry in some GLTF assets is inefficiently indexed; these issues should be fixable in the future by just changing `optimize_indices`.

I would expect that this has significant performance benefits for geometry heavy scenes with LODs on ARM Mali in particular (as the way these GPUs do vertex shading is all sorts of strange), but I don't have that GPU to test. That said, this has benefits for desktop GPUs as well. Testing on my AMD Radeon 7900 GRE, using this geometry heavy scene (3M triangles in view):

![image](https://github.com/user-attachments/assets/e52fbe8e-1ce0-4ea5-a5b1-7971a8c83635)

... I get 1.19 ms GPU time on this PR, vs 1.47 ms GPU time on master. Using "Disable LOD" view on the same camera (17.2M triangles in view), I get 4.20ms GPU time on this PR, vs 4.34ms GPU time on master. This essentially proves that this change does not regress the performance on base level of detail - in fact slightly improving it - and improves this further for LODs; as usual, gains will be dependent on GPU and scene in question (on TPS demo I only see around ~1% improvement in GPU frame time on 7900 GRE for example, but that level has comparatively few LODs and more complex shading).

As before, this change does not change the appearance of the objects in any way - it just changes the vertex order which is not observable. It needs a little bit of extra code to run in import time. A reimport of the model used in this scene takes ~870ms on this PR vs ~860ms on master, so the extra cost should not be noticeable - the most significant cost here is LOD generation (which didn't change), followed by vertex cache optimization (which didn't change, just moved into the new function); the only new cost is remap table generation and remapping which is comparatively quick.

It may make sense to expose the new `optimize_indices` function to scripts now but I'm not sure if this is useful as I'm still fuzzy on the use cases of script-driven imports. Let me know if you'd like me to do this in this PR.

I've tested this on a few glTF models, TPS demo, the MRP from the example with line lists that regressed before, and an MRP from some old issue about blend shapes to make sure this also works with blend shapes (it does). Unsure what else to test on, but should hopefully be correct :)