@lawnjelly 2024-11-04
I haven't looked at the discrete LOD in master, but am I right in thinking it just uses a different index buffer but accessing the same vertices, so the access is always going to be sparse (although with this PR it may improve)?

Is an alternative to store different areas of the VB with duplicate verts (like mipmaps) to better use vertex cache etc? (Trading off using more video memory, but this shouldn't be much in the case of lower LODs.)

(I was aiming to have a look at progressive meshes in 3.x at some point, and was planning on duplicating parts of the VB to form different zones like mipmaps.)