@zeux 2024-11-04
Also tested on a variant of the scene from https://github.com/godotengine/godot/issues/93587 (without artificial 8x duplication, so just 100 models) from this view:

![image](https://github.com/user-attachments/assets/7cf108ea-8b3b-405d-a2e7-cdcf2112974a)

Before: GPU time 0.64ms normal, 1.84ms disable LOD; import time 25.9s
After: GPU time 0.56ms normal, 1.78ms disable LOD; import time 25.9s (~unchanged)

Due to the refactoring, the optimization cost is now isolated and more clearly visible in the profiler:

![image](https://github.com/user-attachments/assets/3a39bbc4-95c6-4cec-8843-13332b64f8b5)